refactor: 상태 객체들 패키지 분리
