+++ src/main/java/blackjack/state/NeedToCompareState.java
@@ -2,6 +2,7 @@ package blackjack.state;
 
 import blackjack.gamer.Dealer;
 import blackjack.gamer.Player;
+import blackjack.state.finalState.DrawState;
 
 public class NeedToCompareState implements State {
