+++ src/main/java/blackjack/state/finalState/LoseState.java
@@ -1,10 +1,10 @@
-package blackjack.state;
+package blackjack.state.finalState;
 
 import blackjack.gamer.Dealer;
 import blackjack.gamer.Player;
+import blackjack.state.State;
 
-public class DrawState implements State, FinalState {
-
+public class LoseState implements State, FinalState {
     @Override
     public void checkState(final Player player, final Dealer dealer) {
         calculateEarnedMoney(player);
@@ -12,6 +12,6 @@ public class DrawState implements State, FinalState {
 
     @Override
     public void calculateEarnedMoney(final Player player) {
-        player.drawGame();
+        player.loseGame();
     }
 }
