+++ src/main/java/blackjack/state/finalState/PushState.java
@@ -1,9 +1,11 @@
-package blackjack.state;
+package blackjack.state.finalState;
 
 import blackjack.gamer.Dealer;
 import blackjack.gamer.Player;
+import blackjack.state.State;
 
-public class LoseState implements State, FinalState {
+public class PushState implements State, FinalState {
+
     @Override
     public void checkState(final Player player, final Dealer dealer) {
         calculateEarnedMoney(player);
@@ -11,6 +13,6 @@ public class LoseState implements State, FinalState {
 
     @Override
     public void calculateEarnedMoney(final Player player) {
-        player.loseGame();
+        player.pushGame();
     }
 }
