+++ src/test/java/blackjack/blackjack/BlackjackTest.java
@@ -12,8 +12,8 @@ import blackjack.cardMachine.CardRandomMachine;
 import blackjack.gamer.Dealer;
 import blackjack.gamer.Player;
 import blackjack.gamer.Players;
-import blackjack.state.LoseState;
-import blackjack.state.PushState;
+import blackjack.state.finalState.LoseState;
+import blackjack.state.finalState.PushState;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
